requirements.txt
langchain-huggingface
langchain-core
langchain-community
langchain-chroma
torch
gradio